Migrated methods from deprecated ResourcesScreenDomainService
